requirements.txt
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
h11==0.14.0
httptools==0.6.1
idna==3.10
joblib==1.4.2
msgpack==1.1.0
numpy==1.24.4
pandas==2.0.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
scikit-learn==1.3.2
scikit-surprise==1.1.4
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
wsproto==1.2.0
